# Roca Maria/SparqlEndpoint/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Documentation",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/map-data": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Returnează datele job-urilor pentru hartă cu filtrare",
                "operationId": "276a5220e4ede3b34c1f2238a8315e49",
                "parameters": [
                    {
                        "name": "dateRange",
                        "in": "query",
                        "description": "Intervalul temporal",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all",
                            "enum": [
                                "all",
                                "lastWeek",
                                "last3Months",
                                "lastYear",
                                "upcoming",
                                "past"
                            ],
                            "example": "lastWeek"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listă de job-uri cu coordonate geografice",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "job": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "http://jobs.example.com/job/123"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "location": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "Cluj-Napoca, Romania"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "coordinates": {
                                                "properties": {
                                                    "lat": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 46.7712
                                                    },
                                                    "lng": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 23.6236
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "jobIRI": {
                                                "type": "string",
                                                "example": "http://jobs.example.com/job/123"
                                            },
                                            "title": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "Full Stack Developer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "job": {
                                            "value": "http://jobs.example.com/job/123"
                                        },
                                        "location": {
                                            "value": "București, Romania"
                                        },
                                        "coordinates": {
                                            "lat": 44.4268,
                                            "lng": 26.1025
                                        },
                                        "jobIRI": "http://jobs.example.com/job/123",
                                        "title": {
                                            "value": "Senior Software Developer"
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Nu s-au găsit job-uri",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No jobs data found"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "error": "No jobs data found"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events-map-data": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Returnează datele evenimentelor pentru hartă cu posibilitate de filtrare",
                "operationId": "245b95a738aedbab9fe4c9c9db2cd64a",
                "parameters": [
                    {
                        "name": "eventType",
                        "in": "query",
                        "description": "Tipul evenimentului pentru filtrare",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "conference"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de evenimente cu coordonate geografice",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "event": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "http://cultural-events.org/event/123"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "location": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "Sibiu, Romania"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "coordinates": {
                                                "properties": {
                                                    "lat": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 45.7983
                                                    },
                                                    "lng": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 24.1253
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "eventIRI": {
                                                "type": "string",
                                                "example": "http://cultural-events.org/event/123"
                                            },
                                            "isFromRomania": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "event": {
                                            "value": "http://cultural-events.org/event/123"
                                        },
                                        "location": {
                                            "value": "Sibiu, Romania"
                                        },
                                        "coordinates": {
                                            "lat": 45.7983,
                                            "lng": 24.1253
                                        },
                                        "eventIRI": "http://cultural-events.org/event/123",
                                        "isFromRomania": true
                                    }
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Nu s-au găsit evenimente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No events data found"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "error": "No events data found"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}